added IEnumerable Extension methods
Fixed a bug in the words() implementation of the TextCorpusReader.
Added two extension methods to IEnumerable: Chunk and NGram.  They read
the stream and create "lists of lists" (as lazily as possible). Chunk
simply creates sublists of the specified length until done.  Ngram
creates ngram sublists of the specified length.